+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Added radio button typified element  
 * Fixed sending parameters using PUT method
 
 1.0.0-rc.34